:star: Basic flow to get certs.
Check if we have certs.
If not run HTTP mode.
Check if we're specifying not to get certs first (maybe it's just
tests).
If not check if we have user credentials to use getting certs.
If not send them through OAuth and save their creds.
Set an interval to check for credentials.
When found, fetch certs from CA and save them.